feat(separated-packages): separando os pacotes e utilizando turbo para gerenciar os pacotes
