+++ .github/common-actions/install/action.yml
@@ -1,0 +1,25 @@
+name: 'install and configure git'
+description: 'sets up node.js and runs install'
+
+runs:
+  using: composite
+  steps:
+    - name: Install dependencies
+      uses: pnpm/action-setup@v4
+
+    - name: Setup Node.js
+      uses: actions/setup-node@v4
+      with:
+        node-version-file: '.nvmrc'
+        registry-url: 'https://registry.npmjs.org'
+        cache: 'pnpm'
+
+    - name: Setup Git User
+      shell: bash
+      run: |
+        git config --global user.email "andrefelipeschulle@icloud.com"
+        git config --global user.name "André Felipe Schulle"
+
+    - name: Install dependencies
+      shell: bash
+      run: pnpm install
