--- .github/workflows/publish.yml
@@ -1,58 +1,0 @@
-name: CI - Publish Eslint Config
-
-on:
-  push:
-    branches:
-      - main
-
-jobs:
-  publish-grp:
-    runs-on: ubuntu-latest
-
-    permissions:
-      contents: read
-      id-token: write
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: NodeJs Environment
-        uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - name: Build Packages
-        run: npm install --force && npm run build
-
-      - name: Set package registry GRP
-        run: npm config set registry https://npm.pkg.github.com
-
-      - name: Github package registry authentication
-        run: npm set //npm.pkg.github.com/:_authToken ${{ secrets.PUBLISH_TOKEN_GRP }}
-
-      - name: Publish Packages on GRP
-        run: cd dist && npm publish --provenance --access public
-
-  publish-npm:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      id-token: write
-
-    needs: [publish-grp]
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: NodeJs Environment
-        uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - name: Build Packages
-        run: npm install --force && npm run build
-
-      - name: Npm registry authentication
-        run: npm set //registry.npmjs.org/:_authToken ${{ secrets.NPM_TOKEN }}
-
-      - name: Publish Packages on NPM
-        run: cd dist && npm publish --provenance --access public
