--- .vscode/settings.json
@@ -1,6 +1,0 @@
-{
-  "typescript.tsdk": "node_modules/typescript/lib",
-  "editor.formatOnSave": true,
-  "exportall.config.folderListener": ["/src"],
-  "exportall.config.semis": false
-}
