+++ packages/base/package.json
@@ -1,0 +1,39 @@
+{
+	"name": "@brainylab/eslint-config-base",
+	"version": "3.0.0",
+	"description": "ESLint base config",
+	"keywords": [
+		"brainylab",
+		"eslint",
+		"config",
+		"base"
+	],
+	"license": "MIT",
+	"author": "André Felipe Schulle <andrefelipeschulle@icloud.com>",
+	"type": "module",
+	"main": "./src/index.ts",
+	"files": [
+		"dist",
+		"LICENSE"
+	],
+	"scripts": {
+		"build": "rimraf dist && tsup --dts",
+		"clean": "rimraf dist .turbo",
+		"prepack": "clean-package",
+		"postpack": "clean-package restore",
+		"typecheck": "tsc --noEmit"
+	},
+	"dependencies": {
+		"@eslint/js": "9.15.0",
+		"eslint": "9.15.0",
+		"eslint-config-prettier": "9.1.0",
+		"eslint-define-config": "2.1.0",
+		"eslint-plugin-import-x": "4.4.3",
+		"globals": "15.12.0",
+		"typescript-eslint": "8.16.0"
+	},
+	"devDependencies": {
+		"clean-package": "2.2.0"
+	},
+	"clean-package": "../clean-package.config.json"
+}
