+++ packages/base/src/index.ts
@@ -1,0 +1,162 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
+import pluginJs from '@eslint/js';
+import globals from 'globals';
+import tsEslint from 'typescript-eslint';
+import eslintPluginImportX from 'eslint-plugin-import-x';
+// @ts-ignore
+import eslintConfigPrettier from 'eslint-config-prettier';
+
+import type { FlatESLintConfig } from 'eslint-define-config';
+
+const typeScriptExtensions = ['.ts', '.tsx', 'd.ts'] as const;
+
+const allExtensions = [...typeScriptExtensions, '.js', '.jsx'] as const;
+
+const typescript = tsEslint.config({
+	extends: [...tsEslint.configs.recommended],
+	languageOptions: {
+		parserOptions: {
+			parser: tsEslint.parser,
+			project: './tsconfig.json',
+			ecmaVersion: 'latest',
+			sourceType: 'module',
+			ecmaFeatures: {
+				jsx: true,
+			},
+		},
+	},
+	settings: {
+		'import-x/extensions': allExtensions,
+		'import-x/external-module-folders': ['node_modules', 'node_modules/@types'],
+		'import-x/parsers': {
+			'@typescript-eslint/parser': [...typeScriptExtensions, '.cts', '.mts'],
+		},
+		'import-x/resolver': {
+			node: {
+				extensions: allExtensions,
+			},
+		},
+	},
+	files: ['**/*.{ts,tsx}'],
+	ignores: [
+		'**/node_modules/**/*',
+		'**/build/**/*',
+		'**/coverage/**/*',
+		'**/dist/*',
+		'**/.next/**/*',
+	],
+	rules: {
+		'@typescript-eslint/no-explicit-any': ['warn', { ignoreRestArgs: true }],
+		'@typescript-eslint/no-unused-vars': 'error',
+		'@typescript-eslint/consistent-type-exports': 'error',
+		'@typescript-eslint/consistent-type-imports': [
+			'error',
+			{ fixStyle: 'separate-type-imports' },
+		],
+	},
+});
+
+export const base = [
+	...typescript,
+	{
+		files: ['**/*.{js,cjs,mjs,jsx,mjsx,ts,tsx,mtsx}', '**/*.d.ts'],
+		ignores: [
+			'**/node_modules/**/*',
+			'**/build/**/*',
+			'**/coverage/**/*',
+			'**/dist/*',
+			'**/.next/**/*',
+			'**/*.d.ts',
+			'eslint.config.js',
+		],
+		plugins: {
+			import: eslintPluginImportX,
+		},
+		languageOptions: {
+			parserOptions: {
+				ecmaVersion: 'latest',
+				sourceType: 'module',
+			},
+			globals: {
+				...globals.es2021,
+			},
+		},
+		rules: {
+			...pluginJs.configs.recommended.rules,
+			'no-unused-vars': 'off',
+			'import-x/named': 'off',
+			'import/order': [
+				'error',
+				{
+					'newlines-between': 'always',
+					groups: [
+						'builtin',
+						'external',
+						'internal',
+						'parent',
+						'sibling',
+						'index',
+						'object',
+						'type',
+					],
+					pathGroups: [{ group: 'internal', pattern: '{{@,~}/,#}**' }],
+					pathGroupsExcludedImportTypes: ['type'],
+				},
+			],
+			'eslint-comments/no-unlimited-disable': 'off',
+			'import/no-duplicates': 'off',
+			'unused-imports/no-unused-vars': 'off',
+		},
+	},
+	{
+		name: 'no default exports',
+		rules: {
+			'no-restricted-syntax': [
+				'error',
+				{
+					selector: 'ExportDefaultDeclaration',
+					message: 'Prefer named exports',
+				},
+			],
+		},
+	},
+	{
+		name: 'no default exports exceptions',
+		files: [
+			'src/app/**/{page,layout,template,loading,error,robots,sitemaps}.tsx',
+			'src/app/**/{robots,sitemap}.ts',
+			'*.config.{ts,js,mjs,cjs}',
+			'**/*.config.{ts,js,mjs,cjs}',
+			'*.d.ts',
+			'**/*.d.ts',
+		],
+		rules: {
+			'no-restricted-syntax': ['off', { selector: 'ExportDefaultDeclaration' }],
+		},
+	},
+	{
+		files: ['**/*.{test,spec}.ts?(x)'],
+		rules: {
+			'no-unused-expressions': 'off',
+		},
+	},
+	{
+		files: ['**/*.?([cm])js', '**/*.cjs'],
+		rules: {
+			'@typescript-eslint/no-require-imports': 'off',
+			'@typescript-eslint/no-var-requires': 'off',
+		},
+	},
+	{
+		files: ['**/*.d.ts'],
+		rules: {
+			'no-restricted-syntax': [
+				'error',
+				'ForInStatement',
+				'LabeledStatement',
+				'WithStatement',
+			],
+		},
+	},
+	eslintConfigPrettier,
+] as FlatESLintConfig[];
