+++ packages/base/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"baseUrl": "."
+	},
+	"include": ["src", "tsup.config.ts"]
+}
