+++ packages/base/tsup.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'tsup';
+
+export default defineConfig({
+	entry: ['src/index.ts'],
+	clean: true,
+	target: 'es2019',
+	format: ['esm'],
+});
