+++ packages/clean-package.config.json
@@ -1,0 +1,9 @@
+{
+	"remove": ["devDependencies"],
+	"replace": {
+		"exports": "dist/index.js",
+		"main": "dist/index.js",
+		"module": "dist/index.js",
+		"types": "dist/index.d.ts"
+	}
+}
