--- src/presets/index.ts
@@ -1,7 +1,0 @@
-import { base } from './base';
-import { typescript } from './typescript';
-import { node } from './node';
-import { react } from './react';
-import { next } from './next';
-
-export const presets = { base, typescript, node, react, next };
