--- tsup.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from 'tsup';
-
-export default defineConfig({
-  entry: ['./src/index.ts'],
-  format: ['esm'],
-  target: 'node18.18',
-  dts: true,
-});
